`registration/register.py`: Rename to `registration/algorithms.py`
This will help to distinguish the actual registration algorithms from the other registration functions (entry-points, labeling functions, etc.)
