+++ spinalcordtoolbox/registration/algorithms.py
