+++ spinalcordtoolbox/scripts/sct_register_multimodal.py
@@ -38,7 +38,7 @@ from copy import deepcopy
 import numpy as np
 
 from spinalcordtoolbox.reports.qc import generate_qc
-from spinalcordtoolbox.registration.register import Paramreg, ParamregMultiStep
+from spinalcordtoolbox.registration.algorithms import Paramreg, ParamregMultiStep
 from spinalcordtoolbox.utils.shell import SCTArgumentParser, Metavar, ActionCreateFolder, list_type, display_viewer_syntax
 from spinalcordtoolbox.utils.sys import init_sct, printv, set_loglevel
 from spinalcordtoolbox.utils.fs import extract_fname
